Add files via upload
DISCLAMER - ONLY FOR DEVELOPER USE! 
rename the file to stagesync.py to use it!

StageSync - Verbose Mode

This version of the StageSync file has been configured in verbose mode, ideal for development and debugging purposes. In this mode, the code includes numerous detailed logs that provide an exhaustive trace of the operations performed. This can be extremely useful for identifying and resolving issues, but may be intrusive in production environments.

Use this version only when detailed monitoring of system behavior is required.